app/src/test/java/software/yesaya/sajo/statistics/StatisticsViewModelTest.kt
package software.yesaya.sajo.statistics

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import kotlinx.coroutines.ExperimentalCoroutinesApi
import org.hamcrest.CoreMatchers
import org.junit.Assert
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import software.yesaya.sajo.MainCoroutineRule
import software.yesaya.sajo.data.source.FakeTestRepository
import software.yesaya.sajo.getOrAwaitValue
import software.yesaya.sajo.tasks.ui.statistics.StatisticsViewModel

@ExperimentalCoroutinesApi
class StatisticsViewModelTest {

    // Executes each task synchronously using Architecture Components.
    @get:
    Rule
    var instantExecutorRule = InstantTaskExecutorRule()

    // Set the main coroutines dispatcher for unit testing.
    @ExperimentalCoroutinesApi
    @get:Rule
    var mainCoroutineRule = MainCoroutineRule()

    // Subject under test
    private lateinit var statisticsViewModel: StatisticsViewModel

    // Use a fake repository to be injected into the view model.
    private lateinit var tasksRepository: FakeTestRepository

    @Before
    fun setupStatisticsViewModel() {
        // Initialise the repository with no tasks.
        tasksRepository = FakeTestRepository()

        statisticsViewModel = StatisticsViewModel(tasksRepository)
    }

    @Test
    fun loadTasks_loading() {
        mainCoroutineRule.pauseDispatcher()

        statisticsViewModel.refresh()

        Assert.assertThat(
            statisticsViewModel.dataLoading.getOrAwaitValue(),
            CoreMatchers.`is`(true)
        )

        mainCoroutineRule.resumeDispatcher()

        Assert.assertThat(
            statisticsViewModel.dataLoading.getOrAwaitValue(),
            CoreMatchers.`is`(false)
        )
    }

    @Test
    fun loadStatisticsWhenTasksAreUnavailable_callErrorToDisplay() {
        // Make the repository return errors.
        tasksRepository.setReturnError(true)
        statisticsViewModel.refresh()

        // Then empty and error are true (which triggers an error message to be shown).
        Assert.assertThat(statisticsViewModel.empty.getOrAwaitValue(), CoreMatchers.`is`(true))
        Assert.assertThat(statisticsViewModel.error.getOrAwaitValue(), CoreMatchers.`is`(true))
    }
}